% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRDataTable.R
\name{CRDataTable}
\alias{CRDataTable}
\title{R6 Class representing concentration-response data belonging to a compound}
\description{
R6 Class representing concentration-response data belonging to a compound

R6 Class representing concentration-response data belonging to a compound
}
\details{
Data belongig to a compound has a source file with a seperator, compound name, data and a protocol.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{source_file}}{is the path to source *.csv-file.}

\item{\code{output}}{folder name.}

\item{\code{sep}}{seperator for *.csv-file.}

\item{\code{compound}}{name of compound.}

\item{\code{dt}}{raw data in form of a data.table.}

\item{\code{key_process}}{counter.}

\item{\code{solvent_control}}{wording}

\item{\code{renormalized}}{table}

\item{\code{grid_size}}{to use for curves}

\item{\code{extrapolation}}{factor (for concentration range)}

\item{\code{bg_normalized}}{table}

\item{\code{ctl_normalized}}{marker}

\item{\code{num_trt_ok}}{requirements treatment groups}

\item{\code{num_ctl_rpl_ok}}{requirements control replicates}

\item{\code{cut_off_appl}}{boolean whether custom cut off is applied}

\item{\code{min_num_trt}}{number of treatment groups}

\item{\code{min_num_ctl_rpl}}{number of ctl repl}

\item{\code{yaxis_lim}}{limit for plots}

\item{\code{protocol}}{vector of characters.}

\item{\code{crc_dt}}{boolean}

\item{\code{bmcs}}{data.frame of the bmcs.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{output}}{folder name.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CRDataTable-new}{\code{CRDataTable$new()}}
\item \href{#method-CRDataTable-get_doses}{\code{CRDataTable$get_doses()}}
\item \href{#method-CRDataTable-get_endpoints}{\code{CRDataTable$get_endpoints()}}
\item \href{#method-CRDataTable-get_experiments}{\code{CRDataTable$get_experiments()}}
\item \href{#method-CRDataTable-add_quotient_parameter}{\code{CRDataTable$add_quotient_parameter()}}
\item \href{#method-CRDataTable-save_dt}{\code{CRDataTable$save_dt()}}
\item \href{#method-CRDataTable-custom_cut_off}{\code{CRDataTable$custom_cut_off()}}
\item \href{#method-CRDataTable-conditions}{\code{CRDataTable$conditions()}}
\item \href{#method-CRDataTable-control_replicates}{\code{CRDataTable$control_replicates()}}
\item \href{#method-CRDataTable-background_subtraction}{\code{CRDataTable$background_subtraction()}}
\item \href{#method-CRDataTable-control_normalization}{\code{CRDataTable$control_normalization()}}
\item \href{#method-CRDataTable-renormalize}{\code{CRDataTable$renormalize()}}
\item \href{#method-CRDataTable-to_get}{\code{CRDataTable$to_get()}}
\item \href{#method-CRDataTable-to_getV}{\code{CRDataTable$to_getV()}}
\item \href{#method-CRDataTable-no_effect}{\code{CRDataTable$no_effect()}}
\item \href{#method-CRDataTable-fit_dt}{\code{CRDataTable$fit_dt()}}
\item \href{#method-CRDataTable-bmc_calc}{\code{CRDataTable$bmc_calc()}}
\item \href{#method-CRDataTable-plotc}{\code{CRDataTable$plotc()}}
\item \href{#method-CRDataTable-classification_plot}{\code{CRDataTable$classification_plot()}}
\item \href{#method-CRDataTable-run}{\code{CRDataTable$run()}}
\item \href{#method-CRDataTable-clone}{\code{CRDataTable$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CRDataTable-new"></a>}}
\if{latex}{\out{\hypertarget{method-CRDataTable-new}{}}}
\subsection{Method \code{new()}}{
Defining the average function used in this CRDataTable.


A generic helper function to apply filter on dt according m. Mainly used in methods to ensure data quality.


A generic helper function to add new column given by quotiens according counter_col and denominator_col. Mainly used in methods to generate endpoints induced by quotients.


Normalization, Additive or Mulitplicative


Function for Curve-Fitting that relies on the drc-package.


Function for Curve-Fitting that relies on the drc-package.


Create a new data belongig to a compound object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRDataTable$new(
  source_file = NA,
  ddata,
  sep = ",",
  compound = NA,
  general_setting = NA
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{source_file}}{path to source *.csv-file.}

\item{\code{ddata}}{data.table of data, when data of source_file is available direct in that form}

\item{\code{sep}}{seperator for *.csv-file.}

\item{\code{compound}}{compound name}

\item{\code{general_setting}}{data.frame with general settings}

\item{\code{endpt_conc_table}}{data.table of with columns (Endpoint, Reference Concentration).}

\item{\code{composition}}{binary arithmetic operation, additive or multiplicative}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{CRDataTable} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CRDataTable-get_doses"></a>}}
\if{latex}{\out{\hypertarget{method-CRDataTable-get_doses}{}}}
\subsection{Method \code{get_doses()}}{
Get concentrations.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRDataTable$get_doses()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
List of concentrations in the \code{CRDataTable} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CRDataTable-get_endpoints"></a>}}
\if{latex}{\out{\hypertarget{method-CRDataTable-get_endpoints}{}}}
\subsection{Method \code{get_endpoints()}}{
Get endpoints.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRDataTable$get_endpoints()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
List of endpoints in the \code{CRDataTable} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CRDataTable-get_experiments"></a>}}
\if{latex}{\out{\hypertarget{method-CRDataTable-get_experiments}{}}}
\subsection{Method \code{get_experiments()}}{
Get experiments.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRDataTable$get_experiments()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
List of experiments in the \code{CRDataTable} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CRDataTable-add_quotient_parameter"></a>}}
\if{latex}{\out{\hypertarget{method-CRDataTable-add_quotient_parameter}{}}}
\subsection{Method \code{add_quotient_parameter()}}{
Add new endpoints by quotiens.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRDataTable$add_quotient_parameter(quotients = NA)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CRDataTable-save_dt"></a>}}
\if{latex}{\out{\hypertarget{method-CRDataTable-save_dt}{}}}
\subsection{Method \code{save_dt()}}{
Export current state of dt as csv-file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRDataTable$save_dt(foldername)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CRDataTable-custom_cut_off"></a>}}
\if{latex}{\out{\hypertarget{method-CRDataTable-custom_cut_off}{}}}
\subsection{Method \code{custom_cut_off()}}{
Apply custom cut off on dt according to cutoff_table.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRDataTable$custom_cut_off(cutoff_table = NA)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CRDataTable-conditions"></a>}}
\if{latex}{\out{\hypertarget{method-CRDataTable-conditions}{}}}
\subsection{Method \code{conditions()}}{
Ensure data quality according number of conditions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRDataTable$conditions()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CRDataTable-control_replicates"></a>}}
\if{latex}{\out{\hypertarget{method-CRDataTable-control_replicates}{}}}
\subsection{Method \code{control_replicates()}}{
Ensure data quality according number of control (technical) replicates.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRDataTable$control_replicates()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CRDataTable-background_subtraction"></a>}}
\if{latex}{\out{\hypertarget{method-CRDataTable-background_subtraction}{}}}
\subsection{Method \code{background_subtraction()}}{
Apply background normalization according to the background_table.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRDataTable$background_subtraction(background_table = NA)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CRDataTable-control_normalization"></a>}}
\if{latex}{\out{\hypertarget{method-CRDataTable-control_normalization}{}}}
\subsection{Method \code{control_normalization()}}{
Apply control normalizaiton with respect to dr_endpoints_table.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRDataTable$control_normalization(dr_endpoints_table = NA)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CRDataTable-renormalize"></a>}}
\if{latex}{\out{\hypertarget{method-CRDataTable-renormalize}{}}}
\subsection{Method \code{renormalize()}}{
Apply renormalization according to drc models drc_models.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRDataTable$renormalize(
  experiments = c("merged", "single"),
  endpoint,
  drc_models
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CRDataTable-to_get"></a>}}
\if{latex}{\out{\hypertarget{method-CRDataTable-to_get}{}}}
\subsection{Method \code{to_get()}}{
Get data / model for endpoint with seperation of data regarding all plate (merged) or single plates (single)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRDataTable$to_get(
  sth = c("data", "model"),
  experiments = c("merged", "single"),
  for_endpoint
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sth}}{select data or model.}

\item{\code{experiments}}{merged or single option.}

\item{\code{for_endpoint}}{character or list of endpoints. If not restricted, for all endpoints will be printed.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Named list of the models (list('model' = ..., 'drcObject' = ...)) of the best fit. The names are the plate ids (IUF: Experiment_IDs) for experiments option single, and 'merged' for experiments option merged.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CRDataTable-to_getV"></a>}}
\if{latex}{\out{\hypertarget{method-CRDataTable-to_getV}{}}}
\subsection{Method \code{to_getV()}}{
Get data / model for endpoint with seperation of data regarding all plate (merged) or single plates (single)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRDataTable$to_getV(
  sth = c("data", "model"),
  experiments = c("merged"),
  for_endpoint
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sth}}{select data or model.}

\item{\code{experiments}}{merged or single option.}

\item{\code{for_endpoint}}{character or list of endpoints. If not restricted, for all endpoints will be printed.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Named list of the models (list('model' = ..., 'drcObject' = ...)) of the best fit. The names are the plate ids (IUF: Experiment_IDs) for experiments option single, and 'merged' for experiments option merged.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CRDataTable-no_effect"></a>}}
\if{latex}{\out{\hypertarget{method-CRDataTable-no_effect}{}}}
\subsection{Method \code{no_effect()}}{
Check if there is no-effect for given endpoint, data regarding given doses.
Therefore the nested models linear-regression and one-paramter-model are used, and the hypothesis is tested via tha likelihood-ratio test
implicitly using Wilks' theorem about asymptotic distribution of the ratio.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRDataTable$no_effect(d, e, regarding = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{regarding}}{list of doses.}

\item{\code{experiments}}{defines the seperation method.}

\item{\code{for_endpint}}{endpoint.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Named list of boolean with names experiments.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CRDataTable-fit_dt"></a>}}
\if{latex}{\out{\hypertarget{method-CRDataTable-fit_dt}{}}}
\subsection{Method \code{fit_dt()}}{
Fit curves to the concentration response data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRDataTable$fit_dt(
  experiments = c("merged", "single"),
  endpoints_table = NA,
  rnrm = "no"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{experiments}}{merged or single option.}

\item{\code{for_endpoint}}{character or list of endpoints. If not restricted, for all endpoints will be printed.}

\item{\code{normal}}{TRUE (default) or FALSE for renormalisation of the data regarding the smallest dose.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Plot of the data regarding to the endpoints (and experiments) with the benchmark-response and benchmark-dose.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CRDataTable-bmc_calc"></a>}}
\if{latex}{\out{\hypertarget{method-CRDataTable-bmc_calc}{}}}
\subsection{Method \code{bmc_calc()}}{
BMC calculation based on confidence bands (confidence interval projection) of the drc Package (in fact it is the delta method)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRDataTable$bmc_calc(method = "ci_pr", plate, for_endpoint, bmr)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{Currently just the ci_pr option.}

\item{\code{bmr}}{double or list of double with the benchmark-response}

\item{\code{best_fit}}{fit object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Plot of the data regarding to the endpoints (and experiments) with the benchmark-response and benchmark-dose.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CRDataTable-plotc"></a>}}
\if{latex}{\out{\hypertarget{method-CRDataTable-plotc}{}}}
\subsection{Method \code{plotc()}}{
Print curves of the endpoints.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRDataTable$plotc(
  experiments = c("merged", "single"),
  for_endpoint = NULL,
  bmr = NA
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{experiments}}{merged or single option.}

\item{\code{for_endpoint}}{character or list of endpoints. If not restricted, for all endpoints will be printed.}

\item{\code{bmr}}{double or list of double with the benchmark-response}

\item{\code{normal}}{TRUE (default) or FALSE for renormalisation of the data regarding the smallest dose.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Plot of the data regarding to the endpoints (and experiments) with the benchmark-response and benchmark-dose.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CRDataTable-classification_plot"></a>}}
\if{latex}{\out{\hypertarget{method-CRDataTable-classification_plot}{}}}
\subsection{Method \code{classification_plot()}}{
Print curves of the endpoints.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRDataTable$classification_plot(
  experiments = c("merged"),
  for_endpoint = NULL,
  specific_vs_unspecific_table = NA,
  classification_rules_table = NA
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{experiments}}{merged or single option.}

\item{\code{for_endpoint}}{character or list of endpoints. If not restricted, for all endpoints will be printed.}

\item{\code{normal}}{TRUE (default) or FALSE for renormalisation of the data regarding the smallest dose.}

\item{\code{bmr}}{double or list of double with the benchmark-response}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Plot of the data regarding to the endpoints (and experiments) with the benchmark-response and benchmark-dose.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CRDataTable-run"></a>}}
\if{latex}{\out{\hypertarget{method-CRDataTable-run}{}}}
\subsection{Method \code{run()}}{
Starts a full run.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRDataTable$run(task_file = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{task}}{xlsx file containing the tasks and parameter.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
1
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CRDataTable-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CRDataTable-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRDataTable$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
