% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dunnettSDP.R
\name{dunnettSDP}
\alias{dunnettSDP}
\title{Dunnett multiple comparison test - Step Down Procedure}
\usage{
dunnettSDP(d)
}
\arguments{
\item{d}{data.table with columns dose and endpoint}
}
\description{
This function allows you to apply Dunnett multiple comparison test - Step Down Procedure.
}
\examples{
dunnettSDP(d)
}
\keyword{dunnett}
